# Conflicts:
#	helm/hpcc/Chart.yaml
#	helm/hpcc/templates/_helpers.tpl
#	helm/hpcc/templates/dafilesrv.yaml
#	helm/hpcc/templates/dali.yaml
#	helm/hpcc/templates/dfuserver.yaml
#	helm/hpcc/templates/eclagent.yaml
#	helm/hpcc/templates/eclccserver.yaml
#	helm/hpcc/templates/eclscheduler.yaml
#	helm/hpcc/templates/esp.yaml
#	helm/hpcc/templates/localroxie.yaml
#	helm/hpcc/templates/roxie.yaml
#	helm/hpcc/templates/sasha.yaml
#	helm/hpcc/templates/thor.yaml
#	version.cmake

+++ dockerfiles/image.sh
@@ -323,7 +323,9 @@ install() {
 }
 
 function cleanup() {
-    rm $RSYNC_TMP_FILE 2> /dev/null || true
+    if [ -v RSYNC_TMP_FILE ]; then
+        rm $RSYNC_TMP_FILE 2> /dev/null || true
+    fi
 }
 
 trap cleanup EXIT
